refactor(fcm): move fcm package under push output plugin
The fcm package is only used in the push output plugin, so move it.
